requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
contourpy==1.3.2
cryptography==46.0.5
cycler==0.12.1
distro==1.9.0
exceptiongroup==1.3.1
fastapi==0.135.1
filelock==3.20.0
fonttools==4.62.1
fsspec==2025.12.0
google-auth==2.49.1
google-genai==1.68.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hub-sdk==0.0.24
idna==3.11
Jinja2==3.1.6
kiwisolver==1.5.0
MarkupSafe==3.0.2
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.11.0.86
opencv-python-headless==4.9.0.80
packaging==26.0
pandas==2.3.3
pillow==12.0.0
psutil==7.2.2
py-cpuinfo==9.0.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-multipart==0.0.22
pytz==2026.1.post1
PyYAML==6.0.3
requests==2.32.5
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
starlette==0.52.1
sympy==1.13.1
tenacity==9.1.4
thop==0.1.1.post2209072238
torch==2.6.0+cu124
torchvision==0.21.0+cu124
tqdm==4.67.3
triton==3.2.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ultralytics==8.1.0
urllib3==2.6.3
uvicorn==0.42.0
websockets==16.0
